+++ spec/repository_manager_spec.rb
@@ -12,7 +12,8 @@ describe Inspector::GitlabManager do
       manager
     end
 
-    [:retrieve_dependencies, :versions, :source_url, :project_by_name].each do |method|
+    [:project_dependencies, :tags, :source_url, :project_by_name,
+     :project_metadata, :project_metadata_version].each do |method|
       it "responds to #{method}" do
         manager.respond_to?(method).should be_true
       end