+++ CHANGELOG.md
@@ -9,6 +9,7 @@ FEATURES:
 * Use a file to store configuration info (defaults to *${HOME}/.chef/kitchen_inspector.rb*)
 * Add option to show remarks (*--remarks*), that now are hidden by default
 * Add support for different Repository Managers (but only Gitlab is supported at this stage)
+* Verifies that a tag on the Repository Manager matches the corresponding metadata's version
 
 FIXES: